Update README.md
Location-based augmented reality is a markerless technique that allows a mobile device to display relevant virtual information about the surroundings on its screen by using GPS and digital compass data. In this study, a mobile application was developed to inform users about buildings belonging to an institution using location-based augmented reality. Unity and Vuforia were used for this purpose. By utilizing the angle of the phone relative to the north and the angles between the corner coordinates of the buildings, the specific building was identified and information about the building was displayed on the phone screen. The application was used within the institution to record the latitude, longitude, and distance information of the buildings. With the recorded data, Decision Trees, Random Forests, Support Vector Machines, and XGBoost algorithms were trained. According to the results obtained, the best performance was achieved by Support Vector Machines, with accuracy, precision, recall, and F1 Score values of 0.90, 0.91, 0.90, and 0.90, respectively. By using machine learning algorithms within the application, building detection was accomplished using only the phone’s latitude, longitude, and distance to the building information, thus eliminating the need for angular calculations for building detection.